move breeding functionality to new context file
this is cleaner to hold data context and breeding context apart.
also remove await from modifyBreedingUpdates as it is not asynchronous.
