+++ frontend/src/main.tsx
@@ -8,6 +8,7 @@ import { BrowserRouter } from "react-router-dom";
 import { WithUserContext } from "@app/user_context";
 import { WithDataContext } from "@app/data_context";
 import { WithMessageContext } from "@app/message_context";
+import { WithBreedingContext } from "@app/breeding_context";
 import { Navigation } from "@app/navigation";
 
 const CSS = sc.createGlobalStyle`
@@ -29,13 +30,15 @@ const Main = (
   <>
     <CssBaseline />
     <BrowserRouter>
-      <WithMessageContext>
-        <WithDataContext>
-          <WithUserContext>
-            <Navigation />
-          </WithUserContext>
-        </WithDataContext>
-      </WithMessageContext>
+      <WithDataContext>
+        <WithUserContext>
+          <WithMessageContext>
+            <WithBreedingContext>
+              <Navigation />
+            </WithBreedingContext>
+          </WithMessageContext>
+        </WithUserContext>
+      </WithDataContext>
     </BrowserRouter>
     <CSS />
   </>
